Rename cluster modules
